% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{annotate_magma}
\alias{annotate_magma}
\title{Create gene annotation file for MAGMA (and install MAGMA if not available)}
\usage{
annotate_magma(gene_matrix, settings = gm_settings)
}
\arguments{
\item{settings}{list with magma installation settings}
}
\description{
Create gene annotation file for MAGMA (and install MAGMA if not available)
}

