% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{check_path}
\alias{check_path}
\title{Checks if path exists}
\usage{
check_path(path_name)
}
\arguments{
\item{path_name}{to check}
}
\value{
None
}
\description{
Checks if path exists. If not, stop execution.
}

