% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_genebasedp.R
\name{compute_genebased_p}
\alias{compute_genebased_p}
\title{Compute gene-based pvalue based on daner summary stats file}
\usage{
compute_genebased_p(summary_file = NULL, output_prefix = NULL,
  annot_prefix = NULL, ref_file = NULL, snpmap_file = NULL,
  magma_executable = NULL, magma_model = "snp-wise=mean")
}
\arguments{
\item{settings}{a list containing all}
}
\value{
filename with gene-based pvalues
}
\description{
Computes gene-based pvalues with magma and saves them to file.
}

