% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_genebasedp.R
\name{compute_genebased_pvalues}
\alias{compute_genebased_pvalues}
\title{Compute gene-based pvalue based on daner summary stats files available}
\usage{
compute_genebased_pvalues(settings = gm_settings)
}
\arguments{
\item{settings}{a list containing all magma-related settings}
}
\description{
Compute gene-based pvalue based on daner summary stats files available
}

