% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_genebasedp.R
\name{compute_genebasedp}
\alias{compute_genebasedp}
\title{Compute gene-based pvalue based on daner summary stats file}
\usage{
compute_genebasedp(summary_file = NULL, output_prefix = NULL,
  annot_prefix = NULL, ref_file = NULL, magma_executable = NULL)
}
\arguments{
\item{settings}{a list containing all}
}
\value{
Data table with one p-value per gene
}
\description{
Compute gene-based pvalue based on daner summary stats file
}

