% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{create_dir}
\alias{create_dir}
\title{Create directory if it does not exist}
\usage{
create_dir(dir_name)
}
\arguments{
\item{dir_name}{directory name to check}
}
\value{
None
}
\description{
Create directory if it does not exist
}

