% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genematrix.R
\name{create_genematrix}
\alias{create_genematrix}
\title{Create and save annotated Gene Matrix}
\usage{
create_genematrix(result_dir = getwd(), settings = gm_settings)
}
\description{
Create and save annotated Gene Matrix
}

