% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtt_functions.R
\name{create_symbol_table}
\alias{create_symbol_table}
\title{Create gene symbol translation table}
\usage{
create_symbol_table(core, value_sep, excel_col = T)
}
\description{
Create gene symbol translation table
}

