% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{download_source}
\alias{download_source}
\title{Download source url if it does not exist in destination path}
\usage{
download_source(source_url, dest_path)
}
\arguments{
\item{source_url}{url to source file to download}

\item{dest_path}{path to save source file in}
}
\value{
None
}
\description{
Checks if data source already exists in destination path.
If not it downloads the data source and saves to destionation path.
}

