% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genematrix.R
\name{generate_genematrix}
\alias{generate_genematrix}
\title{Save annotated Gene Matrix}
\usage{
generate_genematrix(settings = gm_settings)
}
\arguments{
\item{settings}{a named list with user specified settings. Defaults to global defaults in gm_setting
stored in config.R}
}
\description{
Main function to generate a default Gene Matrix and store it in a data.table.
The first time this may take a long time (up to an hour). Will be much quicker due to caching later.
}

