% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R
\name{get_core_matrix}
\alias{get_core_matrix}
\title{Load core gene matrix and create it first if it does not exist}
\usage{
get_core_matrix(settings = gm_settings)
}
\arguments{
\item{settings}{list with settings (defaults to global genematrix settings gm_settings)}
}
\value{
data.table with core gene matrix
}
\description{
A cache wrapper around \code{create_core_matrix()}
}

