% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_magma_osversion}
\alias{get_magma_osversion}
\title{Get compiled magma platform version to download: windows (win), mac (mac osx), static (linux fully static )}
\usage{
get_magma_osversion()
}
\description{
Get compiled magma platform version to download: windows (win), mac (mac osx), static (linux fully static )
}

