% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R
\name{is_par}
\alias{is_par}
\title{Check whether a genomic position is in a PAR regions
Assumes chr as factor 1-22,X,Y,M}
\usage{
is_par(chr, start_pos, end_pos, build = "b37")
}
\description{
Check whether a genomic position is in a PAR regions
Assumes chr as factor 1-22,X,Y,M
}

