% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtt_functions.R
\name{lookup_symbol}
\alias{lookup_symbol}
\title{Lookup gene symbol in gene-translation-table (gtt)}
\usage{
lookup_symbol(symbols, symbol_table)
}
\arguments{
\item{symbols}{character of gene symbols to look up in gtt}

\item{symbol_table}{list containing gene_translation_table (mappging from aliases to official gene name)}
}
\value{
character vector with official gene names
}
\description{
Lookup gene symbol in gene-translation-table (gtt)
}

