% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{merge_annotation_type}
\alias{merge_annotation_type}
\title{Merge processed annotation file to gene matrix}
\usage{
merge_annotation_type(annot_label, process_function, gene_matrix,
  gene_translation_table, settings)
}
\arguments{
\item{annot_label}{label used to name annotation in genematrix columnname}

\item{process_function}{function to process the annotation only accepting settings as parameter}

\item{gene_matrix}{data.table with gene matrix to merge annotation with}

\item{gene_translation_table}{data.table to look up official gene symbols}

\item{settings}{list with global settings}
}
\description{
Merge processed annotation file to gene matrix
}

