% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_commonmindDLPFCdiffexp.R
\name{merge_commonmindDLPFC}
\alias{merge_commonmindDLPFC}
\title{Merge omim phenotypes to gene matrix}
\usage{
merge_commonmindDLPFC(gene_matrix, gene_translation_table,
  settings = gm_settings)
}
\description{
Merge omim phenotypes to gene matrix
}

