% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R, R/functions.R
\name{merge_core_entrez}
\alias{merge_core_entrez}
\title{Merge gencode and hgnc file}
\usage{
merge_core_entrez(core, entrez)

merge_core_entrez(core, entrez)
}
\description{
Merge gencode and hgnc file

Merge gencode and hgnc file
}

