% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_exacpli.R
\name{merge_exacpli}
\alias{merge_exacpli}
\title{Merge exac pli constraint data into gene matrix}
\usage{
merge_exacpli(pli_url, gene_matrix, gene_translation_table,
  gene_bp_dif = 5e+05)
}
\description{
Merge exac pli constraint data into gene matrix
}

