% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gwascatalog.R
\name{merge_gwascatalog}
\alias{merge_gwascatalog}
\title{Merge exac pli constraint data into gene matrix}
\usage{
merge_gwascatalog(gene_matrix, gene_translation_table, settings = gm_settings)
}
\description{
Merge exac pli constraint data into gene matrix
}

