% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_omim.R
\name{merge_omim}
\alias{merge_omim}
\title{Merge omim phenotypes to gene matrix}
\usage{
merge_omim(gene_matrix, gene_translation_table, settings = gm_settings)
}
\description{
Merge omim phenotypes to gene matrix
}

