% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_genebasedp.R
\name{merge_snpnrs}
\alias{merge_snpnrs}
\title{Merge nr of SNPs per gene used in gene-based test}
\usage{
merge_snpnrs(gene_matrix, settings = gm_settings)
}
\arguments{
\item{gene_matrix}{a data.table with gene annotation}

\item{settings}{a list containing all magma-related settings}
}
\value{
a data.table with nr of SNPs added
}
\description{
Merge nr of SNPs per gene used in gene-based test
}

