% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R
\name{merge_url_links}
\alias{merge_url_links}
\title{Merge url links to common resources}
\usage{
merge_url_links(core)
}
\arguments{
\item{core}{data.table with gene definition}
}
\value{
data.table with core gene matrix including url links to relevant resources
}
\description{
Merge url links to common resources
}

