% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_commonmindDLPFCdiffexp.R
\name{process_commonmindDLPFC}
\alias{process_commonmindDLPFC}
\title{Process omim morbidmap.txt file (source data.omim.org/downloads, registration/license required)}
\usage{
process_commonmindDLPFC(settings = gm_settings)
}
\description{
Process omim morbidmap.txt file (source data.omim.org/downloads, registration/license required)
}

