% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R, R/functions.R
\name{process_gencodefile}
\alias{process_gencodefile}
\title{Preprocess gencode file}
\usage{
process_gencodefile(gencode_path)

process_gencodefile(gencode_path)
}
\arguments{
\item{gencode_path}{file path of downloaded .gtf.gz file}

\item{gencode_path}{file path of downloaded .gtf.gz file}
}
\value{
processed gencode table as \code{data.table}

processed gencode table as \code{data.table}
}
\description{
Filter gencode file to include only genes on chromosome 1-22,X,Y,M and reformat before returning as \code{data.table}

Filter gencode file to include only genes on chromosome 1-22,X,Y,M and reformat before returning as \code{data.table}
}
\details{
Gencode files are downloaded from \link{http://www.gencodegenes.org/releases/grch37_mapped_releases.html}

Assumed input format .gtf.gz:
\itemize{
 \item{1) }{chromosome name chr\{1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,X,Y,M\} or GRC accession}
 \item{2) }{annotation source \{ENSEMBL,HAVANA\}}
 \item{3) }{feature type \{gene,transcript,exon,CDS,UTR,start_codon,stop_codon,Selenocysteine\}}
 \item{4) }{genomic start location integer-value (1-based)}
 \item{5) }{genomic end location integer-value}
 \item{6) }{score (not used)}
 \item{7) }{genomic strand \{+,-\}}
 \item{8) }{genomic phase (for CDS features) \{0,1,2,.\}}
 \item{9) }{additional information as key-value pairs}
}
First, the entries are filtered on \code{feature_type == 'gene'} and \code{status == 'KNOWN'}.
This mostly excludes transcripts. The 'chr' prefix is removed from chromosome values and any
chromosomes other than 1-2,X,Y,M are removed. The resulting chromosome values are cast into
an ordered factor (ordering: 1-22,X,Y,M). Then additional columns are extracted from the
key,value pairs in the info column. Any genes with gene_types in \code{c('misc_RNA','snoRNA','snRNA')}
are removed. Finally the redundant columns score, phase, and info are removed and a new column ensembl_gene_id
is created from gene_id that does not contain subnumbering (i.e. id is x instead of x.y).
The resulting file still contains duplicate gene names, but these will be removed after the merge with
the canonical hgnc data.

Gencode files are downloaded from \link{http://www.gencodegenes.org/releases/grch37_mapped_releases.html}

Assumed input format .gtf.gz:
\itemize{
 \item{1) }{chromosome name chr\{1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,X,Y,M\} or GRC accession}
 \item{2) }{annotation source \{ENSEMBL,HAVANA\}}
 \item{3) }{feature type \{gene,transcript,exon,CDS,UTR,start_codon,stop_codon,Selenocysteine\}}
 \item{4) }{genomic start location integer-value (1-based)}
 \item{5) }{genomic end location integer-value}
 \item{6) }{score (not used)}
 \item{7) }{genomic strand \{+,-\}}
 \item{8) }{genomic phase (for CDS features) \{0,1,2,.\}}
 \item{9) }{additional information as key-value pairs}
}
First, the entries are filtered on \code{feature_type == 'gene'} and \code{status == 'KNOWN'}.
This mostly excludes transcripts. The 'chr' prefix is removed from chromosome values and any
chromosomes other than 1-2,X,Y,M are removed. The resulting chromosome values are cast into
an ordered factor (ordering: 1-22,X,Y,M). Then additional columns are extracted from the
key,value pairs in the info column. Any genes with gene_types in \code{c('misc_RNA','snoRNA','snRNA')}
are removed. Finally the redundant columns score, phase, and info are removed and a new column ensembl_gene_id
is created from gene_id that does not contain subnumbering (i.e. id is x instead of x.y).
The resulting file still contains duplicate gene names, but these will be removed after the merge with
the canonical hgnc data.
}
\examples{
\dontrun{
gencode_data <- process_gencodefile(gencode_path)
}
\dontrun{
gencode_data <- process_gencodefile(gencode_path)
}
}

