% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_omim.R
\name{process_omim}
\alias{process_omim}
\title{Process omim morbidmap.txt file (source data.omim.org/downloads, registration/license required)}
\usage{
process_omim(settings = gm_settings)
}
\description{
Process omim morbidmap.txt file (source data.omim.org/downloads, registration/license required)
}

