% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genematrix.R
\name{publish_genematrix}
\alias{publish_genematrix}
\title{Save customized gene matrix based on settings}
\usage{
publish_genematrix(gene_matrix, settings)
}
\arguments{
\item{gene_matrix}{a data.table with gene annotation to save}

\item{settings}{a list including which columns to save and destination path}
}
\value{
None
}
\description{
Publish gene matrix in .csv format including user-set columns
}

