% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtt_functions.R
\name{uniform_genenames}
\alias{uniform_genenames}
\title{Standardize gene names to upper case and remove punctuations (other than .)}
\usage{
uniform_genenames(genenames)
}
\arguments{
\item{genenames}{vector of genenames}
}
\value{
a vector of sanitized gene names (capital letters, most punctuations removed)
}
\description{
Standardize gene names to upper case and remove punctuations (other than .)
}

