% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitity_functions.R
\name{annotate_magma}
\alias{annotate_magma}
\title{Install stand-alone magma if not available yet}
\usage{
annotate_magma(settings = gm_settings)
}
\description{
Install stand-alone magma if not available yet
}

