% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtt_functions.R
\name{get_symbol_table}
\alias{get_symbol_table}
\title{Load gene name translation table and create if it does not exist}
\usage{
get_symbol_table(core_matrix, settings,
  excel_col = settings$add_excel_collisions)
}
\description{
Load gene name translation table and create if it does not exist
}

