% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{gm_settings}
\alias{gm_settings}
\title{Global settings available after loading GeneMatrix package}
\format{An object of class \code{list} of length 16.}
\usage{
gm_settings
}
\description{
Global settings available after loading GeneMatrix package
}
\keyword{datasets}

