% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R
\name{process_entrezfile}
\alias{process_entrezfile}
\title{Preprocess entrez file as downloaded from
ftp.ncbi.nih.gov/gene/DATA/GENE_INFO/Mammalia/Homo_sapiens.gene_info.gz > gene info file from ncbi}
\usage{
process_entrezfile(entrez_path, value_sep)
}
\arguments{
\item{entrez}{gene info file path of downloaded .gz file}
}
\value{
processed gene info table as data.table
}
\description{
Assumed input format .gz:
}
\details{
Processing steps:

1) Change | separator into specified value separator
}
\examples{
\dontrun{
entrez_file <- process_entrez(entrez_path)
}
}

