% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corematrix_functions.R, R/functions.R
\name{process_hgncfile}
\alias{process_hgncfile}
\title{Preprocess hgnc file as downloaded from
http://www.genenames.org/cgi-bin/statistics > Complete hgnc_dataset (.txt)}
\usage{
process_hgncfile(hgnc_path, value_sep)

process_hgncfile(hgnc_path, value_sep)
}
\arguments{
\item{hgnc}{file path of downloaded .txt file}

\item{hgnc}{file path of downloaded .txt file}
}
\value{
processed hgnc table as data.table

processed hgnc table as data.table
}
\description{
Assumed input format .txt:

Assumed input format .txt:
}
\details{
Processing steps:

1) Restrict to relevant columns (as specified in the function)

2) Remove genes that are withdrawn

3) Add alias column which merges gene symbols from alias_symbol and prev_symbol in a single column

4) Change '|' value separater into ';'

Processing steps:

1) Restrict to relevant columns (as specified in the function)

2) Remove genes that are withdrawn

3) Add alias column which merges gene symbols from alias_symbol and prev_symbol in a single column

4) Change '|' value separater into ';'
}
\examples{
\dontrun{
hgnc_data <- process_hgnc(hgnc_path)
}
\dontrun{
hgnc_data <- process_hgnc(hgnc_path)
}
}

